Python/DiscordBot/MatsB/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
discord-py-interactions==5.3.1
discord-typings==0.5.1
emoji==2.2.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
tomli==2.0.1
typing_extensions==4.5.0
yarl==1.9.2
